animation system fix+port, content creation tools

+++ RuntimeData.cs
@@ -14,10 +14,10 @@ namespace TanksRebirth;
 #pragma warning disable CA2211
 public static class RuntimeData {
     /// <summary>Total memory used by the Garbage Collector.</summary>
-    public static ulong GCMemory => (ulong)GC.GetTotalMemory(false);
-    public static ulong MemoryUsageInBytes;
+    public static long GCMemory => GC.GetTotalMemory(false);
+    public static long MemoryUsageInBytes;
 
-    /// <summary>Currently used physical memory by this application in bytes. Expensive to access to use sparingly.</summary>
+    /// <summary>Currently used physical memory by this application in bytes. Expensive to access, so use sparingly.</summary>
     public static long ProcessMemory {
         get {
             using Process process = Process.GetCurrentProcess();
